Rename actions file and add test to check if checked revisions are cleared when the revisions prop changes
